fix(vcl): use Vary instead of hash_always_miss and cleanup
Using hash_always_miss is a flawed approach because it can cause
a thundering herd to the new backend. (Not to be confused with the
thundering herd to the original backend.) When the hash_always_miss
VCL was written, there was no good alternative though. Now that
Fastly handles Vary superseding correctly, that is a much better
alternative. See also my talk about this at https://vimeo.com/376921144

And to avoid confusion Redirect is now renamed to Static/Redirect, and
hlx_type_redirect to hlx_type_static_redirect.