fix(backends): use a working placeholder backend
the anywhere.run domain is not working (no DNS record, no Fastly service), so Fastly marks the backend as unhealthy. As a temp fix, use the one service we already have. At a later point, we can create a www.anywhere.run service that serves both as documentation and as a canary to catch misconfiguration.