fix(preflight): do not process preflight requests for external embeds
The `embed` request type is quite different from other request types, as it goes to a different host, so the typical method of determining owner, repo, ref does not work. On the other hand, the underlying service does not make use of strains at all, so there is litte point in trying to resolve a strain. This change will skip preflight requests for embeds. A side effect is that the version-picker for helix-embed won't work for users in helix-pages.

fixes #725